fixing the connection
